+++ src/test/java/racingcar/MockRandomNumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar;
 
 import racingcar.domain.random.RandomNumberGenerator;
